+++ backend-golang/pkg/cat/kl.go
@@ -1,0 +1,22 @@
+package cat
+
+import (
+	"fmt"
+
+	"github.com/CC-RMD-EpiBio/gofluttercat/backend-golang/models"
+	"github.com/mederrata/ndvek"
+)
+
+type KLSelector struct {
+	Temperature float64
+}
+
+func (ks KLSelector) NextItem(bs *models.BayesianScorer) *models.Item {
+	abilities, err := ndvek.NewNdArray([]int{1}, []float64{bs.Running.Mean()})
+	if err != nil {
+		panic(err)
+	}
+	fish := bs.Model.FisherInformation(abilities)
+	fmt.Printf("fish: %v\n", fish)
+	return nil
+}
