+++ backend-golang/pkg/cat/variance.go
@@ -1,0 +1,13 @@
+package cat
+
+import (
+	"github.com/CC-RMD-EpiBio/gofluttercat/backend-golang/models"
+)
+
+type VarianceSelector struct {
+	Temperature float64
+}
+
+func (vs VarianceSelector) NextItem(s *models.BayesianScorer) *models.Item {
+	return nil
+}
